feat: Implement de/serialization for the frame tree
It is now possible to create a JSON string representation of the frames for saving  it to a file.
Also one can load a JSON string apply it to an existing tree. This will overwrite the frame transformations and is useful if one has an actual hardware setup which not exactly positioned as the default assumption in the code.